Add build definitions for testing plugins
Add two build definitions (and a plugin configuration option) that
make writing plugin end-to-end tests smoother. They are essentially a
convenience wrapper around `please_repo_e2e_test` that simplify the
process of building the plugin and making it loadable within the test
repo.

This requires a version of Please with the `export` command, and would
benefit from https://github.com/thought-machine/please/pull/3411 (which
allows repos without a top-level `BUILD` file to be buildable), hence
the dependency on Please >= 17.19.1.